requirements.txt
# Decisivis Core Requirements
# Secure dependencies with authentication

# Core ML
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Database
psycopg2-binary==2.9.9
psycopg2-pool==1.1

# Security & Authentication
passlib==1.7.4
argon2-cffi==23.1.0
PyJWT==2.8.0
bcrypt==4.1.1
cryptography==41.0.7

# Caching
redis==5.0.1
upstash-redis==1.0.0

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
sse-starlette==1.8.2

# HTTP
requests==2.31.0
httpx==0.25.2

# Environment
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
mypy==1.7.1